+++ source/model.cpp
@@ -3,7 +3,7 @@
 Model::Model(const Device& device)
 		: device(device), vertexBuffer(nullptr), indexBuffer(nullptr) {}
 
-Model::Model(const Device& device, const std::vector<Vertex>& vertexData, const std::vector<uint16_t> indexData)
+Model::Model(const Device& device, const std::vector<Vertex>& vertexData, const std::vector<uint32_t> indexData)
 		: Model(device) {
 
 	vertexBuffer = new VertexBuffer(device, vertexData);
